src/backend/requirements.txt
json
pydantic
pytest
pandas
markdown
openai
langchain_community
langchain_openai
fastapi
uvicorn